refactor(demo): renamed folder
