+++ projects/demo/src/pages/kit/components/strict-mode-description/strict-mode-api-description.component.ts
