+++ projects/demo/src/pages/kit/components/strict-mode-description/strict-mode-description.template.html
