+++ projects/demo/src/pages/kit/date-range/date-range-mask-doc.component.ts
@@ -11,8 +11,8 @@ import {tuiPure} from '@taiga-ui/cdk';
 import {TuiLinkModule, TuiTextfieldControllerModule} from '@taiga-ui/core';
 import {TuiInputModule} from '@taiga-ui/kit';
 
-import {StrictModeApiDescriptionComponent} from '../components/strict-mode-descriptino/strict-mode-api-description.component';
-import {StrictModeDescriptionComponent} from '../components/strict-mode-descriptino/strict-mode-description.component';
+import {StrictModeApiDescriptionComponent} from '../components/strict-mode-description/strict-mode-api-description.component';
+import {StrictModeDescriptionComponent} from '../components/strict-mode-description/strict-mode-description.component';
 import {DateRangeMaskDocExample1} from './examples/1-date-localization/component';
 import {DateRangeMaskDocExample2} from './examples/2-min-max/component';
 import {DateRangeMaskDocExample3} from './examples/3-min-max-length/component';
