+++ projects/demo/src/pages/kit/date-time/date-time-mask-doc.component.ts
@@ -11,8 +11,8 @@ import {tuiPure} from '@taiga-ui/cdk';
 import {TuiLinkModule, TuiTextfieldControllerModule} from '@taiga-ui/core';
 import {DATE_TIME_SEPARATOR, TuiInputModule} from '@taiga-ui/kit';
 
-import {StrictModeApiDescriptionComponent} from '../components/strict-mode-descriptino/strict-mode-api-description.component';
-import {StrictModeDescriptionComponent} from '../components/strict-mode-descriptino/strict-mode-description.component';
+import {StrictModeApiDescriptionComponent} from '../components/strict-mode-description/strict-mode-api-description.component';
+import {StrictModeDescriptionComponent} from '../components/strict-mode-description/strict-mode-description.component';
 import {DateTimeMaskDocExample1} from './examples/1-date-time-localization/component';
 import {DateTimeMaskDocExample2} from './examples/2-date-time-separator/component';
 import {DateTimeMaskDocExample3} from './examples/3-min-max/component';
